Fixed a bug, in which the spec map would sometimes not get drawn properly:
1) Set to open on text.
2) Do a text search.
3) Return to search spec.
4) Open the map.

Prior to this fix, the map would be displayed off to the upper left. This is what happens when Google Maps creates a map in an undrawn and un-dimensioned element. The element needs to be displayed before the map is created. I fixed it by moving the map creation to the last step, and ensuring that the map container was displayed.